requirements.txt
streamlit
google-genai
python-dotenv
youtube-transcript-api